requirements.txt
fastapi 
uvicorn 
slack_sdk 
requests 
openai 
python-dotenv
